revise files structure, apply style fixes, add .clang-format file
