+++ src/onnxruntime_benchmark/.clang-format
@@ -1,0 +1,65 @@
+IndentWidth: 4
+UseTab: Never
+ColumnLimit: 120
+
+Language: Cpp
+Standard: c++17
+
+AccessModifierOffset: -4
+AlignAfterOpenBracket: Align
+AlignConsecutiveAssignments: false
+AlignConsecutiveDeclarations: false
+AlignConsecutiveMacros: true
+AlignTrailingComments: false
+AllowAllArgumentsOnNextLine: false
+AllowAllConstructorInitializersOnNextLine: false
+AllowAllParametersOfDeclarationOnNextLine: false
+AllowShortBlocksOnASingleLine: Empty
+AllowShortFunctionsOnASingleLine: Empty
+AllowShortIfStatementsOnASingleLine: Never
+AllowShortLambdasOnASingleLine: Empty
+AllowShortLoopsOnASingleLine: false
+AlwaysBreakAfterDefinitionReturnType: None
+AlwaysBreakAfterReturnType: None
+AlwaysBreakBeforeMultilineStrings: false
+AlwaysBreakTemplateDeclarations: Yes
+AllowShortEnumsOnASingleLine: false
+BinPackArguments: false
+BinPackParameters: false
+BreakBeforeBraces: Custom
+BraceWrapping:
+  AfterClass: false
+  BeforeElse: true
+  SplitEmptyFunction: false
+  SplitEmptyRecord: false
+CommentPragmas: '^#'
+Cpp11BracedListStyle: true
+DerivePointerAlignment: false
+FixNamespaceComments: true
+IncludeBlocks: Regroup
+IncludeCategories:
+IncludeBlocks: Regroup
+IncludeCategories:
+  # Current project headers
+  - Regex:           '"([A-Za-z0-9.\/-_])+"'
+    Priority:        1
+  # 3rd party headers with extension
+  - Regex:           '<onnxruntime_cxx_api.h>'
+    Priority:        3
+  - Regex:           '<(gflags|opencv2|nlohmann)\/([A-Za-z0-9.\/-_])+>'
+    Priority:        3
+  # C system headers with extension
+  - Regex:           '<([A-Za-z0-9.\/-_])+\.h>'
+    Priority:        4
+  # C++ standard library headers without extension
+  - Regex:           '<([A-Za-z0-9\/-_])+>'
+    Priority:        5
+IndentCaseLabels: true
+IndentPPDirectives: AfterHash
+PenaltyReturnTypeOnItsOwnLine: 1000000
+SortIncludes: true
+SortUsingDeclarations: true
+SpaceAfterCStyleCast: false
+ForEachMacros:
+  - foreach
+  - FOREACH_CHILD
